# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI_Chatbot API Skeleton
  version: 0.1.0-skeleton
  description: Auto-generated skeleton from workbook; schemas are intentionally coarse.
servers:
- url: http://localhost:8080
paths:
...
  /v1/admin/kb/documents:
    get:
      operationId: kbDocList
      summary: 문서 목록 조회
      description: 문서 목록 조회 기능 제공
      tags:
      - 지식베이스 관리
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        required: false
        schema:
          type: string
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
      - name: to
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Workbook skeleton response (200)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: KB-DOC-LIST
      x-source-row: 25
    post:
      operationId: kbDocUpload
      summary: 문서 업로드
      description: 문서 업로드 기능 제공
      tags:
      - 지식베이스 관리
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: KB-DOC-UPLOAD
      x-source-row: 24
  /v1/admin/kb/documents/{doc_id}/approve:
    post:
      operationId: kbDocApprove
      summary: 문서 승인
      description: 문서 승인 기능 제공
      tags:
      - 지식베이스 관리
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: KB-DOC-APPROVE
      x-source-row: 26
  /v1/admin/kb/documents/{doc_id}/versions/{version}/rollback:
    post:
      operationId: kbDocRollback
      summary: 문서 버전 롤백
      description: 문서 버전 롤백 기능 제공
      tags:
      - 지식베이스 관리
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - ADMIN
      x-program-id: KB-DOC-ROLLBACK
      x-source-row: 27
  /v1/admin/kb/index-operations:
    get:
      operationId: opsKbIndexStatus
      summary: KB indexing operation status
      description: Query ingest/index/vector lifecycle status with failures.
      tags:
      - 지식베이스 관리
      parameters:
      - name: kb_id
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
      - name: to
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Workbook skeleton response (200)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - OPS
      x-program-id: OPS-KB-INDEX-STATUS
      x-source-row: 81
  /v1/admin/kb/reindex:
    post:
      operationId: kbReindexRequest
      summary: 재색인 요청
      description: 재색인 요청 기능 제공
      tags:
      - 지식베이스 관리
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Workbook skeleton response (202)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - OPS
      x-program-id: KB-REINDEX-REQUEST
      x-source-row: 28
  /v1/admin/kb/reindex/{job_id}:
    get:
      operationId: kbReindexStatus
      summary: 재색인 상태 조회
      description: 재색인 상태 조회 기능 제공
      tags:
      - 지식베이스 관리
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Workbook skeleton response (200)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - OPS
      x-program-id: KB-REINDEX-STATUS
      x-source-row: 29
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT